chore: move popup rendering to render-popup.ts
