+++ src/content/popup/popup-container.ts
@@ -1,0 +1,6 @@
+export function getPopupContainer(): HTMLElement | null {
+  const hostElem = document.getElementById('tenten-ja-window');
+  return hostElem && hostElem.shadowRoot
+    ? hostElem.shadowRoot.querySelector('.container')
+    : null;
+}
