+++ src/content/popup/selected-index.ts
@@ -1,4 +1,4 @@
-import { PopupOptions } from './popup';
+import { PopupOptions } from './render-popup';
 
 export function getSelectedIndex(options: PopupOptions, numEntries: number) {
   return options.copyState.kind !== 'inactive' && numEntries
